chore(server): split sender functions
